restructuring utils.cpp
