+++ hashing.h
@@ -1,0 +1,3 @@
+#include <stdint.h>
+
+uint32_t calc_crc32(const uint8_t *const p, const size_t len);
