+++ time.cpp
@@ -1,0 +1,10 @@
+#include <stdint.h>
+#include <time.h>
+
+uint64_t get_us()
+{
+	struct timespec tv;
+	clock_gettime(CLOCK_REALTIME, &tv);
+
+	return uint64_t(tv.tv_sec) * uint64_t(1000 * 1000) + uint64_t(tv.tv_nsec / 1000);
+}
