+++ time.h
@@ -1,0 +1,3 @@
+#include <stdint.h>
+
+uint64_t get_us();
